move limit switcher into filter row
