+++ ohmg/frontend/svelte/src/components/buttons/PaginationButtons.svelte
@@ -6,13 +6,11 @@
   export let loading = false;
   export let currentOffset = 0;
   export let currentLimit = 100;
-  export let showLimitSwitcher = true;
   export let total = 0;
 
   const locale = document.documentElement.lang || 'en';
   $: formattedTotal = total.toLocaleString(locale);
 
-  let limitOptions = [10, 25, 50, 100];
   // this weirdness is necessary because the Select element seems to convert basic value into objects after it's
   // been initialized :/
   $: useLimit = typeof currentLimit == 'string' ? currentLimit : currentLimit.value;
@@ -31,19 +29,6 @@
     <CaretDoubleLeft />
   </button>
   <div>{currentOffset} - {currentOffset + limitInt < total ? currentOffset + limitInt : total} ({formattedTotal})</div>
-  {#if currentLimit != 0 && showLimitSwitcher}
-    <div style="display:flex; flex-direction:row; align-items:center; margin-left:10px;">
-      <span>show</span>
-      <Select
-        items={limitOptions}
-        bind:value={currentLimit}
-        searchable={false}
-        clearable={false}
-        containerStyles="background:none; border:none; margin-left:6px; padding:0; width:45px; cursor:pointer;"
-        listAutoWidth={false}
-      />
-    </div>
-  {/if}
   <button
     class="is-icon-link"
     title="Go to next page"
